requirements.txt
bs4==0.0.1      
descartes==1.1.0
geopandas==0.6.2           
ipykernel==5.1.3   
keplergl==0.1.2      
matplotlib==3.1.3      
pandas==1.0.1     
python-dotenv==0.12.0
requests==2.22.0     
scipy==1.4.1      
seaborn==0.10.0     
zoopla==1.0.0    